app/requirements.txt
gunicorn
Flask==3.1.0
flask_login==0.6.3
mysql-connector-python
numpy==2.2.5
python-slugify==8.0.4
Werkzeug==3.1.3
bcrypt==4.3.0
python-dotenv